qa_db_project/requirements.txt
# requirements.txt
psycopg2-binary
pytest
requests